vanna-flask-main/requirements.txt
flask
vanna[postgres]
db-dtypes
python-dotenv
chromadb
python-docx==1.1.2
ollama